refactor(slack): drop CLI-only web-search display creep, move slack verbs into adapter
Second cleanup pass over the out-of-channels diff:

- Revert the CLI web-search args display feature (web-search-display.ts,
  format-args-display changes + test). After consolidating the duplicated
  predicate, its only consumer was the CLI transcript renderer — the Slack
  card path uses isWebSearchTool from tool-name-mapping plus its own
  strings, so this was unrelated CLI rendering work riding in a Slack PR.
  Recoverable from branch history if worth re-proposing separately.

- Move getRandomSlackAssistantStatusVerb (single consumer: the Slack
  adapter) out of cli/helpers/thinking-messages.ts into the adapter,
  dropping the unused agentName parameter. thinking-messages.ts reverts
  to main state.
