+++ src/cli/helpers/thinking-messages.ts
@@ -43,12 +43,6 @@ const THINKING_VERBS = Object.freeze([
   "internalizing",
 ] as const);
 
-const SLACK_ASSISTANT_STATUS_VERBS = Object.freeze([
-  "cogitating",
-  "thinking",
-  "processing",
-] as const);
-
 export const SYSTEM_PROMPT_UPGRADE_TIP =
   "Use /system to upgrade to the latest default prompt.";
 
@@ -120,12 +114,6 @@ function getRandomVerb(): string {
 // Get a random thinking verb phrase (e.g., "is thinking", "is processing")
 export function getRandomThinkingVerb(): string {
   return `is ${getRandomVerb()}`;
-}
-
-export function getRandomSlackAssistantStatusVerb(agentName?: string): string {
-  const index = Math.floor(Math.random() * SLACK_ASSISTANT_STATUS_VERBS.length);
-  const verb = `is ${SLACK_ASSISTANT_STATUS_VERBS[index] ?? "thinking"}`;
-  return agentName ? `${agentName} ${verb}...` : `${verb}...`;
 }
 
 // Get a random past tense verb (e.g., "thought", "processed")
